Check lib defs in TypeScript type checks
This is kind of a TDD-style commit related to #16. It will fail in CI
test runs. This patch doesn't fix #16, but it fixes the reason that
invalid type defs were not previously caught on CI. I had explicitly
told TypeScript to _not_ validate the library file whose validation was
the whole point of this test... :facepalm: