Fix flaky memory usage tests
Increase the size of the XML that is supposed to fail with out of memory
error when validating with default options. Also explicitly invoke V8's
GC so that the test runs would perhaps be a bit more stable and wouldn't
sometimes fail and sometimes pass.

This is an attempt to fix pipeline
https://github.com/noppa/xmllint-wasm/actions/runs/5650086152/job/15306262562#step:11:23
which failed with error

> AssertionError [ERR_ASSERTION]: The input did not match the regular expression /out of memory/. Input:
> 'No error'

in Node 18.