Fix uncaught error if worker can't be initialized
Related to #19, this "fixes" an unactionable global uncaught error in
case the xmllint worker could not be initialized for whatever reason.

This patch should not cause runtime application code differences,
because the error that this case created would've already been swallowed
by the earlier error (as per how Promise.finally works on rejected
Promise), but it did create a confusing error to logs / global error
handler.