Skip large file test with Node 12
The large file test for memory consumption limits is very flaky at best,
sometimes the test fails because it didn't consume enough memory for the
first error case to fire, and sometimes it fails because the worker
consumed too much memory. This seems to be a problem especially when
testing with Node v12, so at least for now I'm just going to disable it.