Upgrade libxml2 and emscripten (#25)
* Upgrade libxml2 to v2.13.8

There's something in v2.14.0 and forward that breaks our wasm build.
Needs more investigation, but in order to get an official v5 of this
library out, I'll just upgrade to the latest version before the
problematic 2.14, which is v2.13.8.

* Upgrade github actions dependencies to v4

* Raise minimum Node.js version to 16 and default to 22

Emscripten now seems to emit (at least by default) syntax that is not
supported by Node.js 14. I don't have interest in transpiling that for
version 14, which has been EOL for a long while now. v5 is a major
release anyway, so we can raise the Node.js version requirement a bit
while at it.